Don't poke wrapper->stream directly, use
2003-08-21  Jeffrey Stedfast  <fejj@ximian.com>

	* camel-data-wrapper.c (decode_to_stream): Don't poke
	wrapper->stream directly, use camel_data_wrapper_write_to_stream()
	instead as this simplifies things and makes the imap data wrapper
	implementation Just Work (tm).

	* providers/imap/camel-imap-wrapper.c: changed prototype of
	write_to_stream() to return ssize_t.

svn path=/trunk/; revision=22325

+++ camel/providers/imap/camel-imap-wrapper.c
@@ -43,7 +43,7 @@ static CamelDataWrapperClass *parent_class = NULL;
 /* Returns the class for a CamelDataWrapper */
 #define CDW_CLASS(so) CAMEL_DATA_WRAPPER_CLASS (CAMEL_OBJECT_GET_CLASS(so))
 
-static int write_to_stream (CamelDataWrapper *imap_wrapper, CamelStream *stream);
+static ssize_t write_to_stream (CamelDataWrapper *imap_wrapper, CamelStream *stream);
 
 static void
 camel_imap_wrapper_class_init (CamelImapWrapperClass *camel_imap_wrapper_class)
